add retry function with exponential backoff
