+++ typescript/agentkit/package.json
@@ -41,6 +41,7 @@
   "dependencies": {
     "@across-protocol/app-sdk": "^0.2.0",
     "@alloralabs/allora-sdk": "^0.1.0",
+    "@base-org/account": "^2.2.0",
     "@coinbase/cdp-sdk": "^1.38.0",
     "@coinbase/coinbase-sdk": "^0.20.0",
     "@coinbase/x402": "^0.6.3",
