+++ typescript/agentkit/src/action-providers/cdp/swapUtils.ts
@@ -128,47 +128,3 @@ export async function getTokenDetails(
 
   return { fromTokenDecimals, toTokenDecimals, fromTokenName, toTokenName };
 }
-
-/**
- * Utility function to sleep for a given number of milliseconds
- *
- * @param ms - Number of milliseconds to sleep
- * @returns Promise that resolves after the specified delay
- */
-const sleep = (ms: number): Promise<void> => new Promise(resolve => setTimeout(resolve, ms));
-
-/**
- * Retry function with exponential backoff
- *
- * @param fn - The function to retry
- * @param maxRetries - Maximum number of retries (default: 3)
- * @param baseDelay - Base delay in milliseconds (default: 1000)
- * @returns Promise that resolves with the function result or rejects with the last error
- */
-export async function retryWithExponentialBackoff<T>(
-  fn: () => Promise<T>,
-  maxRetries: number = 3,
-  baseDelay: number = 1000,
-): Promise<T> {
-  let lastError: Error;
-
-  for (let attempt = 0; attempt <= maxRetries; attempt++) {
-    // Wait before each attempt
-    // Calculate delay with exponential backoff: baseDelay * 2^attempt
-    const delay = baseDelay * Math.pow(2, attempt);
-    await sleep(delay);
-
-    try {
-      return await fn();
-    } catch (error) {
-      lastError = error as Error;
-
-      // If this was the last attempt, throw the error
-      if (attempt === maxRetries) {
-        throw lastError;
-      }
-    }
-  }
-
-  throw lastError!;
-}
